==> Language_modelling_using_LSTM/part2_data_filtering_tokanization.py <==
################  DATA EXTRACTION #####################
############### part 2 of  Phrase prediction Algo #####
#Action :load data from Part1 output and filtering and tokanizing the data along with removal of indentations and fuzzy characters  
#Input: text file generated by part 1 (intern1.txt)
#outut: Filtered text file in the form of sequences (intern2.txt)
#
#Author: sumit chandak  (chandaksumit29@gmail.com)
#Date: 1/12/2018
#
##########################################################
import string


def load_file(filename):
	file = open(filename, 'r')
	text = file.read()
	file.close()
	return text

def save_file(lines, filename):
	data = '\n'.join(lines)
	file = open(filename, 'w')
	file.write(data)
	file.close()

def clean_file(data):
    data = data.replace('--',' ')    
    tokens = data.split( )
    table = str.maketrans('', '', string.punctuation)
    tokens = [w.translate(table) for w in tokens]
    tokens = [word for word in tokens if word.isalpha()]
    tokens = [word.lower() for word in tokens]
    return tokens



def main():
    in_filename = 'inter_1.txt'
    data = load_file(in_filename)

    tokens = clean_file(data)

    length = 5 + 1
    sequences = list()
    for i in range(length, len(tokens)):
        
        seq = tokens[i-length:i]
        line = ' '.join(seq)
        sequences.append(line)


    print('Total Sequences: %d' % len(sequences))


    out_filename = 'inter_2.txt'
    save_file(sequences, out_filename)


if __name__ == "__main__":
    main()
